hia: prevent infinite loading
